sendMultiple
