+++ packages/websocket/lib/Connection.js
@@ -22,6 +22,12 @@ class ConnectionWebSocket extends Connection {
     return super.send(element, ...args);
   }
 
+  async sendMultiple(elements) {
+    for (const element of elements) {
+      await this.send(element);
+    }
+  }
+
   // https://tools.ietf.org/html/rfc7395#section-3.6
   footerElement() {
     return new xml.Element("close", {
